requirements.txt
aiohttp==3.9.1
aioschedule==0.5.2
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-mgmt-core==1.4.0
azure-mgmt-subscription==3.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.4
ecdsa==0.18.0
exceptiongroup==1.2.1
fastapi==0.109.0
fonttools==4.50.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
isodate==0.6.1
kiwisolver==1.4.5
Mako==1.3.1
MarkupSafe==2.1.4
marshmallow==3.21.0
matplotlib==3.8.4
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
passlib==1.7.4
phonenumbers==8.13.28
pillow==10.2.0
pip==24.0
pip-system-certs==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
pypng==0.20220715.0
pyTelegramBotAPI==4.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==63.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.35.1
tqdm==4.66.2
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.27.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4